requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
opencv-python==4.8.1.78
imutils==0.5.4
numpy==2.0.2
pandas==2.0.3
torch==2.1.0
torchvision==0.16.0
tensorflow==2.18.0
psycopg2-binary==2.9.5
ultralytics==8.0.196
pillow==10.0.1
werkzeug==2.3.7
matplotlib==3.7.2
requests==2.31.0
PyYAML==6.0.1
tqdm==4.65.0
h5py==3.14.0
#Gunicorn